programs: Introduce workaround for kio-fuse
Some programs support accessing network protocols like smb and sftp on
their own, and they advertise support for this via X-KDE-Protocols in
their desktop files. This will cause Dolphin to open said applications
using the URI instead of copying the file to a temporary location before
executing it.

However, when using kio-fuse, this is unnecessary and unwanted. When
using kio-fuse, Dolphin will automatically translate the URI to a
kio-fuse path on the local disk, which will work much more consistently
and does not require re-entering credentials.

To counter-act this, we add a new path to the top of XDG_DATA_DIRS
pointing to a derivation where we've patched mpv and vlc to avoid this
issue.

TODO: Find a solution that can integrate better with the current system.
It would be nice to find a reliable way to override the current-system
symlink instead.