programs: Improve kio-fuse workaround
The previous workaround worked, but because it put a Nix store path in a
system environment variable, changes to the desktop files would not take
effect as easily. This workaround creates a new derivation that goes
into the current system directly, obviating the need for a more invasive
hack elsewhere, while still avoiding the need to rebuild vlc or mpv.